+++ tests/testthat/test-expr_anova_parametric.R
@@ -686,3 +686,21 @@ testthat::test_that(
     )
   }
 )
+
+
+# dataframe -----------------------------------------------------------
+
+testthat::test_that(
+  desc = "dataframe",
+  code = {
+    testthat::expect_is(
+      statsExpressions::expr_anova_parametric(
+        data = mtcars,
+        x = cyl,
+        y = wt,
+        output = "dataframe"
+      ),
+      "tbl_df"
+    )
+  }
+)