+++ tests/testthat/test-expr_t_nonparametric.R
@@ -147,3 +147,21 @@ testthat::test_that(
     testthat::expect_identical(using_function2, results2)
   }
 )
+
+
+# dataframe -----------------------------------------------------------
+
+testthat::test_that(
+  desc = "dataframe",
+  code = {
+    testthat::expect_is(
+      statsExpressions::expr_t_nonparametric(
+        data = dplyr::filter(movies_long, genre == "Action" | genre == "Drama"),
+        x = "genre",
+        y = rating,
+        output = "dataframe"
+      ),
+      "tbl_df"
+    )
+  }
+)